+++ validator-api/src/contract_cache/mod.rs
@@ -4,6 +4,7 @@
 use crate::nymd_client::Client;
 use ::time::OffsetDateTime;
 use anyhow::Result;
+use mixnet_contract_common::families::FamilyHead;
 use mixnet_contract_common::{
     mixnode::MixNodeDetails, reward_params::RewardingParams, GatewayBond, IdentityKey, Interval,
     MixId, MixNodeBond, RewardedSetNodeStatus,
@@ -61,6 +62,8 @@ struct ValidatorCacheInner {
 
     current_reward_params: Cache<Option<RewardingParams>>,
     current_interval: Cache<Option<Interval>>,
+
+    mix_to_family: Cache<Vec<(IdentityKey, FamilyHead)>>,
 }
 
 fn current_unix_timestamp() -> i64 {
@@ -163,6 +166,8 @@ impl<C> ValidatorCacheRefresher<C> {
         let mixnodes = self.nymd_client.get_mixnodes().await?;
         let gateways = self.nymd_client.get_gateways().await?;
 
+        let mix_to_family = self.nymd_client.get_all_family_members().await?;
+
         let rewarded_set_map = self.get_rewarded_set_map().await;
 
         let (rewarded_set, active_set) =
@@ -182,6 +187,7 @@ impl<C> ValidatorCacheRefresher<C> {
                 active_set,
                 rewarding_params,
                 current_interval,
+                mix_to_family,
             )
             .await;
 
@@ -255,6 +261,7 @@ impl ValidatorCache {
         })
     }
 
+    #[allow(clippy::too_many_arguments)]
     async fn update_cache(
         &self,
         mixnodes: Vec<MixNodeDetails>,
@@ -263,6 +270,7 @@ impl ValidatorCache {
         active_set: Vec<MixNodeDetails>,
         rewarding_params: RewardingParams,
         current_interval: Interval,
+        mix_to_family: Vec<(IdentityKey, FamilyHead)>,
     ) {
         match time::timeout(Duration::from_millis(100), self.inner.write()).await {
             Ok(mut cache) => {
@@ -272,6 +280,7 @@ impl ValidatorCache {
                 cache.active_set.update(active_set);
                 cache.current_reward_params.update(Some(rewarding_params));
                 cache.current_interval.update(Some(current_interval));
+                cache.mix_to_family.update(mix_to_family)
             }
             Err(e) => {
                 error!("{}", e);
@@ -445,6 +454,16 @@ impl ValidatorCache {
         }
     }
 
+    pub async fn mix_to_family(&self) -> Cache<Vec<(IdentityKey, FamilyHead)>> {
+        match time::timeout(Duration::from_millis(100), self.inner.read()).await {
+            Ok(cache) => cache.mix_to_family.clone(),
+            Err(e) => {
+                error!("{}", e);
+                Cache::new(Vec::new())
+            }
+        }
+    }
+
     pub(crate) async fn interval_reward_params(&self) -> Cache<Option<RewardingParams>> {
         match time::timeout(Duration::from_millis(100), self.inner.read()).await {
             Ok(cache) => cache.current_reward_params.clone(),
@@ -519,6 +538,7 @@ impl ValidatorCacheInner {
             gateways_blacklist: Cache::default(),
             current_interval: Cache::default(),
             current_reward_params: Cache::default(),
+            mix_to_family: Cache::default(),
         }
     }
 }